{"types":"C","runs":45}

+++ problem15.java
@@ -5,7 +5,7 @@ public class problem15 extends ConsoleProgram {
 	public void run() {
 		int firstNum = readInt("First number is: ");
 		int secondNum = readInt("Second number is: ");
-		double average = (firstNum + secondNum) / 2;
+		int average = (int) ((double) (firstNum + secondNum) / 2);
 		println("Arithmetic average is: " + average);		
 	}
 }
